file-processing-service/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-bigquery
google-auth
pydantic
python-dateutil
